[EAD-15] Moved StringUtil to a public util package
