+++ core/src/main/java/com/exadel/etoolbox/anydiff/util/StringUtil.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.exadel.etoolbox.anydiff.comparison;
+package com.exadel.etoolbox.anydiff.util;
 
 import com.exadel.etoolbox.anydiff.Constants;
 import lombok.AccessLevel;
@@ -31,16 +31,17 @@ import java.util.List;
 
 /**
  * Contains utility methods for string manipulation
+ * <u>Note</u>: This class is not a part of public API and is subject to change. You should not use it directly
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-class StringUtil {
+public class StringUtil {
 
     /**
      * Escapes the specified string to be safely displayed in HTML
      * @param value String to escape
      * @return Escaped string
      */
-    static String escape(String value) {
+    public static String escape(String value) {
         return StringUtils.defaultString(value)
                 .replace(Constants.TAG_OPEN, "&lt;")
                 .replace(Constants.TAG_CLOSE, "&gt;");
@@ -51,7 +52,7 @@ class StringUtil {
      * @param value Collection of strings to analyze
      * @return Number of leading spaces
      */
-    static int getIndent(List<? extends CharSequence> value) {
+    public static int getIndent(List<? extends CharSequence> value) {
         if (CollectionUtils.isEmpty(value)) {
             return 0;
         }
@@ -71,7 +72,7 @@ class StringUtil {
      * @param value String to analyze
      * @return Number of leading spaces
      */
-    static int getIndent(CharSequence value) {
+    public static int getIndent(CharSequence value) {
         if (StringUtils.isEmpty(value)) {
             return 0;
         }
@@ -89,7 +90,7 @@ class StringUtil {
      * @param candidates Substrings to search for
      * @return A {@link Pair} object containing the position of the selected substring and the substring itself
      */
-    static Pair<Integer, String> getNearestSubstring(CharSequence value, String... candidates) {
+    public static Pair<Integer, String> getNearestSubstring(CharSequence value, String... candidates) {
         int lastPosition = value.length();
         String lastCandidate = null;
         for (String candidate : candidates) {
@@ -108,7 +109,7 @@ class StringUtil {
      * @param entries Substrings to remove
      * @return A string with all occurrences of the specified substrings removed
      */
-    static String removeAll(String value, String... entries) {
+    public static String removeAll(String value, String... entries) {
         if (StringUtils.isEmpty(value) || ArrayUtils.isEmpty(entries)) {
             return value;
         }
@@ -130,7 +131,7 @@ class StringUtil {
      * @param length The maximum length of each chunk
      * @return A non-null list of chunks
      */
-    static List<String> splitByLength(CharSequence value, int length) {
+    public static List<String> splitByLength(CharSequence value, int length) {
         if (length <= 0) {
             return Collections.emptyList();
         }
@@ -149,7 +150,7 @@ class StringUtil {
      * @param value The string to split
      * @return A non-null list of lines
      */
-    static List<String> splitByNewline(CharSequence value) {
+    public static List<String> splitByNewline(CharSequence value) {
         if (StringUtils.isEmpty(value)) {
             return Collections.emptyList();
         }
@@ -166,7 +167,7 @@ class StringUtil {
      * @param limit The maximum length of the resulting string
      * @return A truncated string
      */
-    static String truncateMiddle(CharSequence value, int limit) {
+    public static String truncateMiddle(CharSequence value, int limit) {
         return truncateMiddle(value, limit, Constants.ELLIPSIS);
     }
 
@@ -178,7 +179,7 @@ class StringUtil {
      * @param ellipsis The ellipsis string to use
      * @return A truncated string
      */
-    static String truncateMiddle(CharSequence value, int limit, String ellipsis) {
+    public static String truncateMiddle(CharSequence value, int limit, String ellipsis) {
         if (value == null) {
             return StringUtils.EMPTY;
         }
@@ -198,7 +199,7 @@ class StringUtil {
      * @param limit The maximum length of the resulting string
      * @return A truncated string
      */
-    static String truncateRight(CharSequence value, int limit) {
+    public static String truncateRight(CharSequence value, int limit) {
         if (value == null) {
             return StringUtils.EMPTY;
         }
